requirements.txt
waitress~=3.0.2
flasgger~=0.9.7.1
Flask~=3.1.0
elastic-apm~=6.23.0
requests~=2.32.3
python-dotenv==1.0.1
pymongo==4.8.0
numpy~=2.2.0
transformers~=4.47.0
onnxruntime~=1.20.1
google-generativeai~=0.8.3